Исправление ошибок

+++ TinkoffPaymentClientApi/ResponseEntity/TinkoffNotification.cs
@@ -1,6 +1,7 @@
 ﻿using Newtonsoft.Json;
 using System.Collections.Generic;
 using TinkoffPaymentClientApi.Attributes;
+using TinkoffPaymentClientApi.Enums;
 using TinkoffPaymentClientApi.Helpers;
 
 namespace TinkoffPaymentClientApi.ResponseEntity {
@@ -10,7 +11,7 @@ namespace TinkoffPaymentClientApi.ResponseEntity {
     public string? Pan { get; set; }
     public string? CardId { get; set; }
     public string? RebillId { get; set; }
-    public decimal? Amount { get; set; }
+    public uint? Amount { get; set; }
     public string? PaymentId { get; set; }
     public string? OrderId { get; set; }
     public string? Status { get; set; }
